requirements.txt
aws_requests_auth
black
boto3
isort
jsonschema
pyjwt
nose2
openapi-spec-validator
prance
pycodestyle
pycrypto
pylint
requests
warrant
